changes of the colliding in the right file
